Agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "pydantic-ai",
    "pydantic",
    "grpcio",
    "grpcio-tools",
    "uvicorn",
    "loguru",
    "python-dotenv",
]
